backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3

